excel_analyzer/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pandas==1.3.3
openpyxl==3.0.9
numpy==1.21.2
python-dotenv==0.19.0
requests==2.26.0
pydantic==1.8.2
aiohttp==3.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 